[fix] AuthContextで認証状態を共有し、ログイン後のリダイレクトを修正
- useAuthをReact Context (AuthProvider)に移行
- _layout.tsxのuseAuthとlogin.tsxのuseAuthが同じ状態を共有
- ログイン成功→isAuthenticated変更→useEffect発火→ホームにリダイレクト
- amazon-cognito-identity-jsを削除、Cognito REST APIを直接使用
